refactor and add Link
